* 兼容青龙版本2.12.x * 新增清理功能 /clean * 修复一些user功能的问题

+++ install.sh
@@ -13,22 +13,23 @@ echo -e "当前青龙版本为:${QL_BRANCH}\n"
 install_depend(){
 
     echo -e "\n1.开始安装所需依赖\n"
-    echo -e "#机器人所需依赖" >>${QLMainPath}/config/extra.sh
     # 包依赖
     apk add zlib zlib-dev libjpeg-turbo libjpeg-turbo-dev gcc python3-dev libffi-dev musl-dev linux-headers
     # 模块依赖
     pip3 install qrcode==7.3.1 Telethon==1.24.0 requests==2.27.1 Pillow==9.0.0 python-socks==1.2.4 async_timeout==4.0.2 prettytable==3.0.0
 
-    if [ `grep "#机器人所需依赖" ${QLMainPath}/config/extra.sh` ];then
-        echo "已设置重启青龙自动启动机器人"
-    else
-        echo -e "解决重启青龙后，jbot失效问题~"
-        echo "apk add zlib zlib-dev libjpeg-turbo libjpeg-turbo-dev gcc python3-dev libffi-dev musl-dev linux-headers" >>${QLMainPath}/config/extra.sh
-        echo "pip3 install qrcode==7.3.1 Telethon==1.24.0 requests==2.27.1 Pillow==9.0.0 python-socks==1.2.4 async_timeout==4.0.2 prettytable==3.0.0" >>${QLMainPath}/config/extra.sh
-        echo "cd ${QLMainPath}/jbot  && pm2 start ecosystem.config.js" >>${QLMainPath}/config/extra.sh
-        echo "cd ${QLMainPath}/ && pm2 start jbot" >>${QLMainPath}/config/extra.sh
+    if [ -f ${QLMainPath}/config/extra.sh ];then
+        if [ `grep "#机器人所需依赖" ${QLMainPath}/config/extra.sh` ];then
+            echo "已设置重启青龙自动启动机器人"
+        else
+            echo -e "解决重启青龙后，jbot失效问题~"
+            echo -e "#机器人所需依赖" >>${QLMainPath}/config/extra.sh
+            echo "apk add zlib zlib-dev libjpeg-turbo libjpeg-turbo-dev gcc python3-dev libffi-dev musl-dev linux-headers" >>${QLMainPath}/config/extra.sh
+            echo "pip3 install qrcode==7.3.1 Telethon==1.24.0 requests==2.27.1 Pillow==9.0.0 python-socks==1.2.4 async_timeout==4.0.2 prettytable==3.0.0" >>${QLMainPath}/config/extra.sh
+            echo "cd ${QLMainPath}/jbot  && pm2 start ecosystem.config.js" >>${QLMainPath}/config/extra.sh
+            echo "cd ${QLMainPath}/ && pm2 start jbot" >>${QLMainPath}/config/extra.sh
+        fi
     fi
-
 
 }
